Add payment service code snippet

+++ smashclub/src/main/java/com/backendsyndicate/smashclub/payment/core/IPayment.java
@@ -3,6 +3,8 @@ package com.backendsyndicate.smashclub.payment.core;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.ResponseEntity;
 
+import java.math.BigDecimal;
+
 public interface IPayment {
     /*
     * Create => Create Transaction
@@ -10,8 +12,8 @@ public interface IPayment {
     * Refund => Refund Transaction -> Perubahan status
     * */
 
-    public ResponseEntity<Object> createTransaction(Long customerId, double totalPrice, String referenceCode, byte transactionType, HttpServletRequest request);
-    public ResponseEntity<Object> paymentTransaction(String transactionCode, byte paymentMethodId, HttpServletRequest request);
+    public ResponseEntity<Object> createTransaction(Long customerId, BigDecimal totalPrice, String referenceCode, int transactionType, HttpServletRequest request);
+    public ResponseEntity<Object> paymentTransaction(String transactionCode, int paymentMethodId, HttpServletRequest request);
     public ResponseEntity<Object> refundTransaction(String transactionCode, String notes, HttpServletRequest request);
 
     public ResponseEntity<Object> paymentMethodList(HttpServletRequest request);
